# specs/001-headless-financial-api/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Tazco Financial API
  description: |
    Headless financial backend service for credit card management, user scoring,
    and payment simulation. All state changes are published to a message stream
    for downstream consumers.
  version: 1.0.0
  contact:
    name: Tazco Platform Team
servers:
  - url: https://{region}-{project}.cloudfunctions.net/api
    description: Firebase Functions (Production)
    variables:
      region:
        default: us-central1
      project:
        default: tazco-financial
  - url: http://localhost:5001/{project}/{region}/api
    description: Firebase Emulator (Development)
    variables:
      region:
        default: us-central1
      project:
        default: tazco-financial
...
security:
  - bearerAuth: []
paths:
...
  /v1/cards/{cardId}/transactions/purchases:
    post:
      tags: [Transactions]
      summary: Simulate a purchase
      description: |
        Record a purchase on the specified card.
        Fails if amount exceeds available credit.
      operationId: createPurchase
      parameters:
        - $ref: '#/components/parameters/CardId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PurchaseInput'
      responses:
        '201':
          description: Purchase recorded
          headers:
            X-Request-ID:
              $ref: '#/components/headers/X-Request-ID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '402':
          description: Insufficient credit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: INSUFFICIENT_CREDIT
                  message: Purchase amount exceeds available credit
                  details:
                    requestedAmount: 1500.00
                    availableCredit: 1000.00
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Card status changed during processing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/cards/{cardId}/transactions/payments:
    post:
      tags: [Transactions]
      summary: Make a payment
      description: |
        Record a payment on the specified card.
        On-time payments increase score (+10 to +50).
        Late payments decrease score (-20 to -100).
      operationId: createPayment
      parameters:
        - $ref: '#/components/parameters/CardId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentInput'
      responses:
        '201':
          description: Payment recorded
          headers:
            X-Request-ID:
              $ref: '#/components/headers/X-Request-ID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
        '400':
          description: Invalid payment (exceeds balance or zero balance)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Concurrent payment in progress
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/cards/{cardId}/transactions:
    get:
      tags: [Transactions]
      summary: List card transactions
      description: Returns paginated list of purchases and payments for a card.
      operationId: listTransactions
      parameters:
        - $ref: '#/components/parameters/CardId'
        - name: type
          in: query
          schema:
            type: string
            enum: [purchase, payment]
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Transaction list
          headers:
            X-Request-ID:
              $ref: '#/components/headers/X-Request-ID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionsListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ============================================================================
  # ADMIN ENDPOINTS
  # ============================================================================
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Firebase Authentication JWT token

  # ============================================================================
  # PARAMETERS
  # ============================================================================
  parameters:
    CardId:
      name: cardId
      in: path
      required: true
      description: Card identifier
      schema:
        type: string
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Client-generated unique key for request deduplication (valid 24h)
      schema:
        type: string
        maxLength: 64
...
    Cursor:
      name: cursor
      in: query
      description: Pagination cursor from previous response
      schema:
        type: string
    Limit:
      name: limit
      in: query
      description: Number of items per page (default 20, max 100)
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20

  # ============================================================================
  # HEADERS
  # ============================================================================
  headers:
    X-Request-ID:
      description: Unique request identifier for tracing
      schema:
        type: string
        format: uuid

  # ============================================================================
  # RESPONSES
  # ============================================================================
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            token_expired:
              value:
                error:
                  code: TOKEN_EXPIRED
                  message: Authentication token has expired
            invalid_token:
              value:
                error:
                  code: INVALID_TOKEN
                  message: Authentication token is invalid
            account_disabled:
              value:
                error:
                  code: ACCOUNT_DISABLED
                  message: Account has been disabled
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  # ============================================================================
  # SCHEMAS
  # ============================================================================
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Machine-readable error code
            message:
              type: string
              description: Human-readable error message
            details:
              type: object
              additionalProperties: true
              description: Additional error context
    PaginationMeta:
      type: object
      properties:
        nextCursor:
          type: string
          nullable: true
          description: Cursor for next page (null if no more pages)
        hasMore:
          type: boolean
        totalCount:
          type: integer
          description: Total items (may be approximate for large datasets)

    # --------------------------------------------------------------------------
    # Health
    # --------------------------------------------------------------------------
...
    CardSummary:
      type: object
      required: [cardId, type, status, limit, balance, availableCredit]
      properties:
        cardId:
          type: string
        type:
          type: string
        status:
          type: string
          enum: [active, suspended, cancelled]
        limit:
          type: number
          format: double
        balance:
          type: number
          format: double
        availableCredit:
          type: number
          format: double
        minimumPayment:
          type: number
          format: double
        nextDueDate:
          type: string
          format: date-time
        nearLimit:
          type: boolean
          description: True if utilization > 90%
...
    PurchaseInput:
      type: object
      required: [amount]
      properties:
        amount:
          type: number
          format: double
          minimum: 0.01
        merchant:
          type: string
          maxLength: 100
    PaymentInput:
      type: object
      required: [amount]
      properties:
        amount:
          type: number
          format: double
          minimum: 0.01
    TransactionResponse:
      type: object
      required: [transaction]
      properties:
        transaction:
          $ref: '#/components/schemas/Transaction'
        card:
          $ref: '#/components/schemas/CardSummary'
          description: Updated card state
    PaymentResponse:
      allOf:
        - $ref: '#/components/schemas/TransactionResponse'
        - type: object
          properties:
            scoreImpact:
              type: object
              properties:
                previousScore:
                  type: integer
                newScore:
                  type: integer
                delta:
                  type: integer
                reason:
                  type: string
                  enum: [on_time, late]
    Transaction:
      type: object
      required: [transactionId, type, amount, status, timestamp]
      properties:
        transactionId:
          type: string
        type:
          type: string
          enum: [purchase, payment]
        amount:
          type: number
          format: double
        merchant:
          type: string
        status:
          type: string
          enum: [completed, failed]
        timestamp:
          type: string
          format: date-time
        paymentStatus:
          type: string
          enum: [on_time, late]
        scoreImpact:
          type: integer
    TransactionsListResponse:
      type: object
      required: [transactions, pagination]
      properties:
        transactions:
          type: array
          items:
            $ref: '#/components/schemas/Transaction'
        pagination:
          $ref: '#/components/schemas/PaginationMeta'

    # --------------------------------------------------------------------------
    # Admin
    # --------------------------------------------------------------------------
...